sensor-data-streaming-project-main/requirements.txt
Flask
Flask-Cors
Flask-SocketIO
confluent-kafka
joblib
python-dotenv
colorama
scikit-learn
gunicorn